requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.5
daphne==4.2.1
Django==5.2.12
django-allauth==65.15.0
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-decouple==2.1
django-redis==6.0.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
eventlet==0.40.4
exceptiongroup==1.3.1
greenlet==3.3.2
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.3.0
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
stripe==14.4.1
tomli==2.4.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.2
